src/pyproject.toml
[tool.poetry]
authors = ["Planeks"]
description = "Client Onboarding Checklist by Planeks"
name = "smartsheet-checklist"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
django-cors-headers = "^4.4"
django-extensions = "^3.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
python-decouple = "^3.8"
python-dotenv = "^1.0"
requests = "^2.31"
smartsheet-python-sdk = "^3.0"
pycryptodome = "^3.20"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
pytest = "^7.4"
pytest-django = "^4.11.0"
isort = "^6.0"
ruff = "^0.6.7"

[tool.ruff]
line-length = 79
target-version = "py312"
exclude = [
  ".git",
  "__pycache__",
  "*.pyc",
  "*.pyo",
  "*.pyd",
  "env/",
  ".venv/",
  "venv/",
  "build/",
  "dist/",
  "migrations",
  "static/",
  "media/",
  "*.log",
  "db.sqlite3",
  ".pytest_cache/",
  ".mypy_cache/",
  ".cache/",
  "frontend",
]

[tool.ruff.lint]
ignore = [
  "E501",
  "D100",
  "D101",
]
select = [
  "E",    # pycodestyle errors (PEP8)
  "F",    # pyflakes (errors)
  "W",    # pycodestyle warnings
  "I",    # isort (import sorting)
  "D",    # docstring conventions
  "UP",   # pyupgrade (modernization)
  "N",    # pep8-naming
  "B",    # flake8-bugbear (common bugs)
  "ANN",  # type annotations
  "C90",  # mccabe (complexity)
  "PT",   # flake8-builtins
  "ERA",  # eradicate commented out code
  "PL",   # pylint style checks
  "DJ",   # Django-specific rules
]

[tool.ruff.format]
quote-style = "double"

[tool.isort]
profile = "black"
blocked_extensions = [
  "rst", "html", "js", "svg", "txt", "css", "scss", "png", "snap", "tsx",
]
combine_as_imports = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
include_trailing_comma = true
use_parentheses = true
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["migrations", ".git", "__pycache__", "LC_MESSAGES", ".venv"]
src_paths = ["src"]
line_length = 79
multi_line_output = 3
no_lines_before = "LOCALFOLDER"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src"]
python_files = ["test*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = """
    --strict-markers
    --tb=short
    --capture=no
    --durations=5
    --color=yes
    --reuse-db
    --ds=config.settings.dev
"""
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks integration tests",
  "unit: marks unit tests",
]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
